REFACTOR: extract NameHelper
